+++ CHANGELOG.md
@@ -6,6 +6,8 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 ## [Unreleased]
 
 ### Added
+- Added support for `ioctl` calls on BSD platforms
+  ([#478](https://github.com/nix-rust/nix/pull/478))
 - Added struct `TimeSpec`
   ([#475](https://github.com/nix-rust/nix/pull/475))
 - Added complete definitions for all kqueue-related constants on all supported
@@ -24,6 +26,8 @@ This project adheres to [Semantic Versioning](http://semver.org/).
   ([#457](https://github.com/nix-rust/nix/pull/457))
 
 ### Changed
+- Removed the `bad` keyword from the `ioctl!` macro
+  ([#478](https://github.com/nix-rust/nix/pull/478))
 - Changed `TimeVal` into an opaque Newtype
   ([#475](https://github.com/nix-rust/nix/pull/475))
 - `kill`'s signature, defined in `::nix::sys::signal`, changed, so that the